Special Classes (requiring Special Enrollment)
These special classes need additional information from the student. Practically, we only want the student to be able to view these special classes but not enroll in them. As such, there is a link instead of a radio button that allows the student to learn more about the class (including how to get into the class).